requirements.txt
fastapi
uvicorn
python-multipart
scikit-learn
numpy
torch
torchvision
Pillow
